Basic ChatServer setup
