+++ .gitignore
@@ -1,4 +1,6 @@
 .vs
+KINESIS/bin
+KINESIS/obj
 PUZZLEBOX/bin
 PUZZLEBOX/obj
 SKELETON-KING/SKELETON-KING.csproj.user
