+++ KINESIS/IProtocolRequestFactory.cs
@@ -1,0 +1,6 @@
+﻿namespace KINESIS;
+
+public interface IProtocolRequestFactory<T>
+{
+    ProtocolRequest<T>? DecodeProtocolRequest(byte[] buffer, int offset, out int updatedOffset);
+}
