+++ KINESIS/KINESIS.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <SuppressNETCoreSdkPreviewMessage>true</SuppressNETCoreSdkPreviewMessage>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\PUZZLEBOX\PUZZLEBOX.csproj" />
+    <InternalsVisibleTo Include="TRANSMUTANSTEIN" />
+  </ItemGroup>
+
+</Project>
