+++ KINESIS/ProtocolRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace KINESIS;
+
+public abstract class ProtocolRequest<T>
+{
+    public abstract void HandleRequest(IDbContextFactory<BountyContext> dbContextFactory, T subject);
+}
