+++ KINESIS/Usings.cs
@@ -1,0 +1,15 @@
+﻿// Keep this file alphabetized.
+
+#region Microsoft Using Directives
+global using Microsoft.EntityFrameworkCore;
+#endregion
+
+#region PUZZLEBOX Project Using Directives
+global using PUZZLEBOX;
+#endregion
+
+#region System Using Directives
+global using System.Collections.Concurrent;
+global using System.Net;
+global using System.Net.Sockets;
+#endregion
