+++ SKELETON-KING.sln
@@ -9,7 +9,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZORGATH", "ZORGATH\ZORGATH.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TRANSMUTANSTEIN", "TRANSMUTANSTEIN\TRANSMUTANSTEIN.csproj", "{E995B774-0D4A-4046-BEDE-8968639A64B9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PUZZLEBOX", "PUZZLEBOX\PUZZLEBOX.csproj", "{08C73F76-4121-4F89-B6C5-BD6BF7315C7B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PUZZLEBOX", "PUZZLEBOX\PUZZLEBOX.csproj", "{08C73F76-4121-4F89-B6C5-BD6BF7315C7B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KINESIS", "KINESIS\KINESIS.csproj", "{317F93C9-EF94-4BC9-A6BF-0282E0643E34}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +35,10 @@ Global
 		{08C73F76-4121-4F89-B6C5-BD6BF7315C7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{08C73F76-4121-4F89-B6C5-BD6BF7315C7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{08C73F76-4121-4F89-B6C5-BD6BF7315C7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{317F93C9-EF94-4BC9-A6BF-0282E0643E34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{317F93C9-EF94-4BC9-A6BF-0282E0643E34}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{317F93C9-EF94-4BC9-A6BF-0282E0643E34}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{317F93C9-EF94-4BC9-A6BF-0282E0643E34}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
