+++ TRANSMUTANSTEIN/ZORGATH/ChampionsOfNewerthRanksTest.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace TRANSMUTANSTEIN;
+﻿namespace ZORGATH;
 
 [TestClass]
 public class ChampionsOfNewerthRanksTest
